# misc/webApi/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"LanBackups API","description":"ASP.NET Core Web API for LAN Backup service","termsOfService":"None","contact":{"name":"Claudiu Farcas","url":"http://www.asp.net","email":""},"license":{"name":"Free as a bird","url":"http://url.com"}},"basePath":"/","paths":{"/api/BackupConfig":{"get":{"tags":["BackupConfig"],"summary":"retrieve all  backup configurations","operationId":"ApiBackupConfigGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"idx","in":"header","required":false,"type":"string"},{"name":"siz","in":"header","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/PaginatedList[BackupConfiguration,String]"}}},"deprecated":false},"post":{"tags":["BackupConfig"],"summary":"create backup configuration","operationId":"ApiBackupConfigPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":["application/json"],"parameters":[{"name":"backup","in":"body","description":"","required":false,"schema":{"$ref":"#/definitions/BackupConfiguration"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/BackupConfiguration"}},"400":{"description":"Client Error","schema":{"$ref":"#/definitions/BackupConfiguration"}}},"deprecated":false}},"/api/BackupConfig/{id}":{"get":{"tags":["BackupConfig"],"summary":"retrieve backup configuration by ID","operationId":"ApiBackupConfigByIdGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/BackupConfiguration"}},"404":{"description":"Client Error","schema":{"$ref":"#/definitions/BackupConfiguration"}}},"deprecated":false},"put":{"tags":["BackupConfig"],"summary":"Update backup configuration","operationId":"ApiBackupConfigByIdPut","consumes":["application/json","text/json","application/json-patch+json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string"},{"name":"backup","in":"body","description":"","required":false,"schema":{"$ref":"#/definitions/BackupConfiguration"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/BackupConfiguration"}},"400":{"description":"Client Error","schema":{"$ref":"#/definitions/BackupConfiguration"}}},"deprecated":false},"delete":{"tags":["BackupConfig"],"summary":"delete backup configuration by id","operationId":"ApiBackupConfigByIdDelete","consumes":[],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"string"}},"404":{"description":"Client Error","schema":{"type":"string"}}},"deprecated":false}},"/api/BackupConfig/client/{clientid}":{"get":{"tags":["BackupConfig"],"summary":"retrieve backup configuration by ID","operationId":"ApiBackupConfigClientByClientidGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"clientid","in":"path","description":"","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/BackupConfiguration"}}},"404":{"description":"Client Error","schema":{"type":"array","items":{"$ref":"#/definitions/BackupConfiguration"}}}},"deprecated":false}},"/api/Logs":{"get":{"tags":["Logs"],"summary":"retrieves paginated logs from DB","operationId":"ApiLogsGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"idx","in":"header","description":"page index","required":false,"type":"string"},{"name":"siz","in":"header","description":"page size","required":false,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/PaginatedList[BackupLog,DateTime]"}}},"deprecated":false},"post":{"tags":["Logs"],"summary":"Create an entry in the logs","operationId":"ApiLogsPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":["application/json"],"parameters":[{"name":"log","in":"body","description":"","required":false,"schema":{"$ref":"#/definitions/BackupLog"}}],"responses":{"200":{"description":"Success","schema":{"format":"int32","type":"integer"}},"400":{"description":"Client Error","schema":{"format":"int32","type":"integer"}}},"deprecated":false}},"/api/Logs/{id}":{"get":{"tags":["Logs"],"summary":"retrieves alog record by ID","operationId":"ApiLogsByIdGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/BackupLog"}},"404":{"description":"Client Error","schema":{"$ref":"#/definitions/BackupLog"}}},"deprecated":false}},"/api/Logs/client/{clientid}":{"get":{"tags":["Logs"],"summary":"retrieve all logs of a clientIP","operationId":"ApiLogsClientByClientidGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"clientid","in":"path","description":"","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/BackupLog"}}},"404":{"description":"Client Error","schema":{"type":"array","items":{"$ref":"#/definitions/BackupLog"}}}},"deprecated":false}},"/api/Logs/config/{configurationid}":{"get":{"tags":["Logs"],"summary":"retrieves all logs of a specific configuration","operationId":"ApiLogsConfigByConfigurationidGet","consumes":[],"produces":["application/json"],"parameters":[{"name":"configurationid","in":"path","description":"","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"$ref":"#/definitions/BackupLog"}}},"404":{"description":"Client Error","schema":{"type":"array","items":{"$ref":"#/definitions/BackupLog"}}}},"deprecated":false}},"/api/Users/register":{"post":{"tags":["Users"],"operationId":"ApiUsersRegisterPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":["application/json"],"parameters":[{"name":"dtouser","in":"body","required":false,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/IdentityResult"}},"400":{"description":"Client Error","schema":{"$ref":"#/definitions/IdentityResult"}}},"deprecated":false}},"/api/Users/login":{"post":{"tags":["Users"],"operationId":"ApiUsersLoginPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"dtouser","in":"body","required":false,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"Success"}},"deprecated":false}},"/api/Users/pwchange":{"post":{"tags":["Users"],"operationId":"ApiUsersPwchangePost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"dtouser","in":"body","required":false,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"Success"}},"deprecated":false}},"/api/Users/list":{"get":{"tags":["Users"],"operationId":"ApiUsersListGet","consumes":[],"produces":[],"responses":{"200":{"description":"Success"}},"deprecated":false},"post":{"tags":["Users"],"operationId":"ApiUsersListPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"user","in":"body","required":false,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"Success"}},"deprecated":false}}},"definitions":{"PaginatedList[BackupConfiguration,String]":{"type":"object","properties":{"pi":{"format":"int32","type":"integer"},"tp":{"format":"int32","type":"integer"},"recs":{"type":"array","items":{"$ref":"#/definitions/BackupConfiguration"}},"hp":{"type":"boolean","readOnly":true},"hn":{"type":"boolean","readOnly":true}}},"BackupConfiguration":{"required":["clientIP"],"type":"object","properties":{"id":{"type":"string"},"rowVersion":{"format":"byte","type":"string"},"clientIP":{"pattern":"^(\\d{1,3})\\.(\\d{1,3})\\.(\\d{1,3})\\.(\\d{1,3})$","type":"string"},"srcFolder":{"type":"string"},"srcUser":{"type":"string"},"srcPass":{"type":"string"},"destLanFolder":{"type":"string"},"destUser":{"type":"string"},"destPass":{"type":"string"},"isActive":{"type":"boolean"},"crontab":{"type":"string"}}},"PaginatedList[BackupLog,DateTime]":{"type":"object","properties":{"pi":{"format":"int32","type":"integer"},"tp":{"format":"int32","type":"integer"},"recs":{"type":"array","items":{"$ref":"#/definitions/BackupLog"}},"hp":{"type":"boolean","readOnly":true},"hn":{"type":"boolean","readOnly":true}}},"BackupLog":{"type":"object","properties":{"id":{"format":"int32","type":"integer"},"rowVersion":{"format":"byte","type":"string"},"clientIP":{"pattern":"^(\\d{1,3})\\.(\\d{1,3})\\.(\\d{1,3})\\.(\\d{1,3})$","type":"string"},"configurationID":{"type":"string"},"description":{"type":"string"},"logError":{"type":"string"},"status":{"type":"string"},"dateTime":{"format":"date-time","type":"string"}}},"User":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"newPassword":{"type":"string"},"isAdmin":{"type":"boolean"},"succeeded":{"type":"boolean"},"isLockedOut":{"type":"boolean"}}},"IdentityResult":{"type":"object","properties":{"succeeded":{"type":"boolean","readOnly":true},"errors":{"type":"array","items":{"$ref":"#/definitions/IdentityError"},"readOnly":true}}},"IdentityError":{"type":"object","properties":{"code":{"type":"string"},"description":{"type":"string"}}}},"securityDefinitions":{}}