# src/LanBackupAgent/LanBackups API/Models/IdentityError.cs
// Code generated by Microsoft (R) AutoRest Code Generator 0.16.0.0
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.

namespace LanBackupAgent.Models
{
    using System;
    using System.Linq;
    using System.Collections.Generic;
    using Newtonsoft.Json;
    using Microsoft.Rest;
    using Microsoft.Rest.Serialization;

    public partial class IdentityError
    {
        /// <summary>
        /// Initializes a new instance of the IdentityError class.
        /// </summary>
        public IdentityError() { }

        /// <summary>
        /// Initializes a new instance of the IdentityError class.
        /// </summary>
        public IdentityError(string code = default(string), string description = default(string))
        {
            Code = code;
            Description = description;
        }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "code")]
        public string Code { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "description")]
        public string Description { get; set; }

    }
}